+++ R/assess_growth.R
@@ -41,8 +41,9 @@ assess_growth <- function(clean_assess_start, clean_assess_end) {
   id_cols <- c(
     "unique_id", "test_year", "bedscode", "nrc_code", "nrc_desc",
     "county_code", "county_desc", "name",
-    "item_subject_area", "item_desc", "test_grade",
-    "test_subject", "subgroup_code", "subgroup_name",
+    "item_subject_area", "item_desc",
+    "test_grade", "test_subject", "test_grade_string",
+    "subgroup_code", "subgroup_name",
     "is_school", "is_district", "is_multigrade_aggregate"
   )
 
@@ -79,7 +80,7 @@ assess_growth <- function(clean_assess_start, clean_assess_end) {
       "l3_count", "l3_pct", "l4_count", "l4_pct",
       "l2_l4_count", "l2_l4_pct",
       "l3_l4_count", "l3_l4_pct",
-      "mean_scale_score", "prof_percentile", "scale_percentile")
+      "mean_scale_score", "proficiency_percentile", "scale_score_percentile")
   acad_join <- c(id_cols, acad_cols)
 
   #join and rename start