+++ tests/testthat/helper_constants.R
@@ -1,5 +1,3 @@
 # #this gets run before tests
-
-assess_all <- dplyr::bind_rows(assess_13, assess_14, assess_15, assess_16)
-growth_all <- dplyr::bind_rows(assess_growth_13_14, assess_growth_14_15, assess_growth_15_16)
-cohort_growth_all <- dplyr::bind_rows(cohort_growth_13_14, cohort_growth_14_15, cohort_growth_15_16)
+agg_all <- dplyr::bind_rows(agg_2015, agg_2016)
+clean_all <- dplyr::bind_rows(clean_2015, clean_2016)